feat: Sprint Polish — PathPool zero-alloc en 70 renderers, transiciones, haptic feedback, cleanup
- PathPool: 480 Path() por frame eliminadas en 70 renderers Canvas (zero-allocation after first frame)
- remember(blocksWithTasks) en MainScreen para evitar recomputación
- Transiciones de navegación: slide horizontal (pantallas), slide vertical (formularios de bloque)
- Haptic feedback en checkbox de tarea y apertura de lootbox
- Deprecaciones resueltas: dayOfMonth -> day, monthNumber -> month.ordinal
- when exhaustivo sin else redundante en WeekCalendar
- Cast redundante eliminado en BlockFormScreen
- Debug logs eliminados en Navigation.kt y SyncManager.kt (solo se conserva Log.e para errores)

+++ composeApp/src/androidMain/kotlin/com/mnebot/riptide/presentation/components/InputFieldDialogs.android.kt
@@ -108,7 +108,7 @@ actual fun DatePickerDialogWrapper(
 ) {
     val state = rememberDatePickerState(
         initialSelectedDateMillis = run {
-            val days = initial.toEpochDays().toLong()
+            val days = initial.toEpochDays()
             days * 24 * 60 * 60 * 1000L
         }
     )
